rename azure functions packages
